refactor: move schema
