+++ magicgui/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Module for defining the schema of a magicgui widget."""
